requirements.txt
flake8
mypy
pylint
python-dotenv
pyModbusTCP==0.2.0
paho-mqtt==1.6.1
numpy==1.23.3
pandas==1.5.0
SQLalchemy==2.0.7
psycopg2-binary==2.9.5
